+++ src/django/api/constants.py
@@ -256,7 +256,8 @@ class APIV1CommonErrorMessages:
     LOCATION_NOT_FOUND = 'The location with the given id was not found.'
     LOCATION_ID_NOT_VALID = 'The value must be a valid id.'
     NO_ACTIVE_CONTRIBUTION_TO_DISSOCIATE = (
-        'You have no active contribution to this location to dissociate.'
+        'You have no active contribution from an approved list to '
+        'dissociate from this location.'
     )
     FACILITY_LIST_NOT_FOUND = 'The list with the given id was not found.'
     LIST_ALREADY_INACTIVE = 'This list is already inactive.'